# src/config/swagger.yaml
swagger: "2.0"
info:
  description: "Controle Simples de Postagens com Usuários"
  version: "1.0.0"
  title: "posts_control"
basePath: "/api/v1"
consumes:
  - "application/json"
produces:
  - "application/json"
...
definitions:
  user:
    additionalProperties: false
    type: "object"
    properties:
      id:
        type: "integer"
      full_name:
        type: "string"
      user_email:
        type: "string"
      user_password:
        type: "string"
  userCreateRequest:
    allOf:
      - $ref: "#/definitions/user"
    required:
      - full_name
      - user_email
      - user_password
    properties:
      id:
        readOnly: true
  userResponse:
    type: object
    properties:
      id:
        type: integer
      full_name:
        type: string
      user_email:
        type: string
...
paths:
  /user:
    get:
      tags:
        - "Users"
      x-swagger-router-controller: app
      summary: list User
      operationId: listUserHandler
      parameters:
        - name: user_id
          in: query
          type: string
          required: false
      responses:
        200:
          description: "success"
          schema:
            type: array
            items:
              $ref: "#/definitions/userResponse"
    post:
      tags:
        - "Users"
      x-swagger-router-controller: app
      summary: create user
      operationId: createUserHandler
      parameters:
        - in: body
          name: user
          description: create user
          schema:
            $ref: "#/definitions/userCreateRequest"
      responses:
        201:
          description: "created"
          schema:
            $ref: "#/definitions/userResponse"
        500:
          description: Internal server error
          schema:
            type: object
            properties:
              type:
                type: string
              message:
                type : string
              details:
                type : object
    put:
      tags:
        - "Users"
      x-swagger-router-controller: app
      summary: update user
      operationId: updateUserHandler
      parameters:
        - in: body
          name: user
          description: update user
          schema:
            additionalProperties: false
            type: "object"
            required:
              - "id"
            properties:
              id:
                type: "integer"
              full_name:
                type: "string"
              user_email:
                type: "string"
              user_password:
                type: "string"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/userResponse"
        500:
          description: Internal server error
          schema:
            type: object
            properties:
              type:
                type: string
              message:
                type : string
              details:
                type : object
    delete:
      tags:
        - "Users"
      x-swagger-router-controller: app
      summary: delete user
      operationId: deleteUserHandler
      parameters:
        - name: user_id
          in: query
          type: integer
          required: true
      responses:
        204:
          description: "User deleted successfully"
        500:
          description: Internal server error
          schema:
            type: object
            properties:
              type:
                type: string
              message:
                type : string
              details:
                type : object
...